use json encoding for create/edit assignments
canvas is rejecting the existing form-formatted create with a 500 error
and no details as to what's wrong.  some tweaks to the form contents
gets us a 201, but the values we're supplying aren't being used.  rather
than continue to fight canvas, let's just format the request as json,
which has no issues.